% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth_data}
\alias{growth_data}
\title{Demo bacterial growth data}
\format{
\subsection{\code{growth_data}}{

A data frame with 44 rows and 4 columns:
\describe{
\item{minutes}{Minutes elapsed since T0}
\item{condition}{Growth condition; a control and a medium additive}
\item{duplo}{Duplo A or B}
\item{OD600}{The OD600, or Optical Density measured at 600 nm wavelength}
}
}
}
\usage{
growth_data
}
\description{
A small fictitious dataset with duplo's for two conditions of growth medium.
}
\details{
Additional details...
}
\keyword{datasets}
