% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth_curves.R
\name{plot_growth_curves}
\alias{plot_growth_curves}
\title{Plot bacterial growth data}
\usage{
plot_growth_curves(growth_data, add_smoother = FALSE)
}
\arguments{
\item{growth_data}{A tibble of data frame in long format, similar to the growth_data demo table of this package.}

\item{add_smoother}{A logical indicating whether to add a smoother to the plot.}
}
\value{
Nothing
}
\description{
Plots bacterial growth curves with duplos and multiple conditions.
}
\details{
Additional details...
}
\examples{
plot_growth_curves(growth_data)
plot_growth_curves(growth_data, add_smoother = TRUE)
}
